Simplified cli_config classes with std::optional
