+++ libmamba/include/mamba/api/configuration.hpp
@@ -13,6 +13,7 @@
 #include "mamba/core/output.hpp"
 #include "mamba/core/util.hpp"
 #include "mamba/api/constants.hpp"
+#include "mamba/api/configuration_impl.hpp"
 
 #include <yaml-cpp/yaml.h>
 
@@ -36,369 +37,12 @@
         exit(0);                                                                                   \
     }
 
-namespace YAML
-{
-    template <>
-    struct convert<mamba::VerificationLevel>
-    {
-        static Node encode(const mamba::VerificationLevel& rhs)
-        {
-            if (rhs == mamba::VerificationLevel::kDisabled)
-            {
-                return Node("disabled");
-            }
-            else if (rhs == mamba::VerificationLevel::kWarn)
-            {
-                return Node("warn");
-            }
-            else if (rhs == mamba::VerificationLevel::kEnabled)
-            {
-                return Node("enabled");
-            }
-            else
-            {
-                return Node();
-            }
-        }
-
-        static bool decode(const Node& node, mamba::VerificationLevel& rhs)
-        {
-            if (!node.IsScalar())
-            {
-                return false;
-            }
-
-            auto str = node.as<std::string>();
-
-            if (str == "enabled")
-            {
-                rhs = mamba::VerificationLevel::kEnabled;
-            }
-            else if (str == "warn")
-            {
-                rhs = mamba::VerificationLevel::kWarn;
-            }
-            else if (str == "disabled")
-            {
-                rhs = mamba::VerificationLevel::kDisabled;
-            }
-            else
-            {
-                throw std::runtime_error(
-                    "Invalid 'VerificationLevel', should be in {'enabled', 'warn', 'disabled'}");
-            }
-
-            return true;
-        }
-    };
-
-    template <>
-    struct convert<mamba::ChannelPriority>
-    {
-        static Node encode(const mamba::ChannelPriority& rhs)
-        {
-            if (rhs == mamba::ChannelPriority::kStrict)
-            {
-                return Node("strict");
-            }
-            else if (rhs == mamba::ChannelPriority::kFlexible)
-            {
-                return Node("flexible");
-            }
-            else if (rhs == mamba::ChannelPriority::kDisabled)
-            {
-                return Node("disabled");
-            }
-            else
-            {
-                return Node();
-            }
-        }
-
-        static bool decode(const Node& node, mamba::ChannelPriority& rhs)
-        {
-            if (!node.IsScalar())
-            {
-                return false;
-            }
-
-            auto str = node.as<std::string>();
-
-            if (str == "strict")
-            {
-                rhs = mamba::ChannelPriority::kStrict;
-            }
-            else if ((str == "flexible") || (str == "true"))
-            {
-                rhs = mamba::ChannelPriority::kFlexible;
-            }
-            else if (str == "disabled")
-            {
-                rhs = mamba::ChannelPriority::kDisabled;
-            }
-            else
-            {
-                return false;
-            }
-
-            return true;
-        }
-    };
-
-    template <>
-    struct convert<fs::path>
-    {
-        static Node encode(const fs::path& rhs)
-        {
-            return Node(rhs.string());
-        }
-
-        static bool decode(const Node& node, fs::path& rhs)
-        {
-            if (!node.IsScalar())
-            {
-                return false;
-            }
-
-            rhs = fs::path(node.as<std::string>());
-            return true;
-        }
-    };
-
-    template <>
-    struct convert<spdlog::level::level_enum>
-    {
-        static Node encode(const spdlog::level::level_enum& rhs)
-        {
-            using namespace spdlog::level;
-
-            return Node(to_string_view(rhs).data());
-        }
-
-        static bool decode(const Node& node, spdlog::level::level_enum& rhs)
-        {
-            using namespace spdlog::level;
-
-            auto name = node.as<std::string>();
-            auto it = std::find(std::begin(level_string_views), std::end(level_string_views), name);
-            if (it != std::end(level_string_views))
-            {
-                rhs = static_cast<level_enum>(it - std::begin(level_string_views));
-                return true;
-            }
-
-            LOG_ERROR
-                << "Invalid log level, should be in {'critical', 'error', 'warning', 'info', 'debug', 'trace', 'off'} but is '"
-                << name << "'";
-            return false;
-        }
-    };
-}
-
-
 namespace mamba
 {
     class ConfigurableInterface;
 
     namespace detail
     {
-        template <class T>
-        struct cli_config
-        {
-            using storage_type = T;
-
-            cli_config(const storage_type& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return true;
-            };
-            T value()
-            {
-                return m_value;
-            };
-
-            storage_type m_value;
-        };
-
-
-        template <>
-        struct cli_config<bool>
-        {
-            using storage_type = int;
-
-            cli_config(const int& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return m_value != 0;
-            };
-            bool value()
-            {
-                if (m_value == 1)
-                {
-                    return true;
-                }
-                else if (m_value == -1)
-                {
-                    return false;
-                }
-                throw std::runtime_error(
-                    "Invalid boolean storage, should be {'-1': false, '0': undefined, '1': true}");
-            };
-
-            int m_value = 0;
-        };
-
-
-        template <>
-        struct cli_config<std::vector<std::string>>
-        {
-            using value_type = std::vector<std::string>;
-            using storage_type = value_type;
-
-            cli_config(const storage_type& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            value_type value()
-            {
-                return m_value;
-            };
-
-            storage_type m_value = {};
-        };
-
-
-        template <>
-        struct cli_config<std::string>
-        {
-            using storage_type = std::string;
-
-            cli_config(const std::string& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            std::string value()
-            {
-                return m_value;
-            };
-
-            std::string m_value = "";
-        };
-
-
-        template <>
-        struct cli_config<fs::path>
-        {
-            using storage_type = fs::path;
-
-            cli_config(const fs::path& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            fs::path value()
-            {
-                return m_value;
-            };
-
-            fs::path m_value = "";
-        };
-
-
-        template <>
-        struct cli_config<std::size_t>
-        {
-            using storage_type = std::ptrdiff_t;
-
-            cli_config(const storage_type& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return m_value > -1;
-            };
-            std::size_t value()
-            {
-                return m_value;
-            };
-
-            storage_type m_value = -1;
-        };
-
-
-        template <>
-        struct cli_config<VerificationLevel>
-        {
-            using storage_type = std::string;
-
-            cli_config(const std::string& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            VerificationLevel value()
-            {
-                return YAML::Node(m_value).as<VerificationLevel>();
-            };
-
-            std::string m_value = "";
-        };
-
-        template <>
-        struct cli_config<ChannelPriority>
-        {
-            using storage_type = std::string;
-
-            cli_config(const std::string& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            ChannelPriority value()
-            {
-                return YAML::Node(m_value).as<ChannelPriority>();
-            };
-
-            std::string m_value = "";
-        };
-
-        template <>
-        struct cli_config<spdlog::level::level_enum>
-        {
-            using value_type = spdlog::level::level_enum;
-            using storage_type = std::string;
-
-            cli_config(const storage_type& value)
-                : m_value(value){};
-
-            bool defined()
-            {
-                return !m_value.empty();
-            };
-            value_type value()
-            {
-                return YAML::Node(m_value).as<spdlog::level::level_enum>();
-            };
-
-            storage_type m_value = "";
-        };
-
         bool has_config_name(const std::string& file);
 
         bool is_config_file(const fs::path& path);
@@ -420,124 +64,6 @@ namespace mamba
         void print_seq_node(YAML::Emitter&, YAML::Node value, YAML::Node source, bool show_source);
         void print_map_node(YAML::Emitter&, YAML::Node value, YAML::Node source, bool show_source);
     }
-
-    namespace detail
-    {
-        template <class T>
-        struct Source
-        {
-            static std::vector<std::string> default_value(const T&)
-            {
-                return std::vector<std::string>({ "default" });
-            };
-
-            static void merge(const std::map<std::string, T>& values,
-                              const std::vector<std::string>& sources,
-                              T& value,
-                              std::vector<std::string>& source);
-
-            static T deserialize(const std::string& value);
-
-            static bool is_sequence();
-        };
-
-        template <class T>
-        struct Source<std::vector<T>>
-        {
-            static std::vector<std::string> default_value(const std::vector<T>& init)
-            {
-                return std::vector<std::string>(init.size(), "default");
-            };
-
-            static void merge(const std::map<std::string, std::vector<T>>& values,
-                              const std::vector<std::string>& sources,
-                              std::vector<T>& value,
-                              std::vector<std::string>& source);
-
-            static std::vector<T> deserialize(const std::string& value);
-
-            static bool is_sequence();
-        };
-
-        template <class T>
-        void Source<T>::merge(const std::map<std::string, T>& values,
-                              const std::vector<std::string>& sources,
-                              T& value,
-                              std::vector<std::string>& source)
-        {
-            bool value_set = false;
-            source.clear();
-
-            for (auto& s : sources)
-            {
-                auto& val = values.at(s);
-
-                if (!value_set)
-                {
-                    value = val;
-                    value_set = true;
-                }
-                source.push_back(s);
-            }
-        }
-
-        template <class T>
-        T Source<T>::deserialize(const std::string& value)
-        {
-            if (value.empty())
-            {
-                return YAML::Node("").as<T>();
-            }
-            else
-            {
-                return YAML::Load(value).as<T>();
-            }
-        }
-
-        template <class T>
-        bool Source<T>::is_sequence()
-        {
-            return false;
-        }
-
-        template <class T>
-        void Source<std::vector<T>>::merge(const std::map<std::string, std::vector<T>>& values,
-                                           const std::vector<std::string>& sources,
-                                           std::vector<T>& value,
-                                           std::vector<std::string>& source)
-        {
-            value.clear();
-            source.clear();
-
-            for (auto& s : sources)
-            {
-                auto& vec = values.at(s);
-                for (auto& v : vec)
-                {
-                    auto find_v = std::find(value.begin(), value.end(), v);
-
-                    if (find_v == value.end())
-                    {
-                        value.push_back(v);
-                        source.push_back(s);
-                    }
-                }
-            }
-        }
-
-        template <class T>
-        std::vector<T> Source<std::vector<T>>::deserialize(const std::string& value)
-        {
-            return YAML::Load("[" + value + "]").as<std::vector<T>>();
-        }
-
-        template <class T>
-        bool Source<std::vector<T>>::is_sequence()
-        {
-            return true;
-        }
-    }
-
 
     enum class ConfigurationLevel
     {
